⚡ pre-compile Regex for QueryPlan::Literal to improve executor speed, fix macOS build

+++ src/lib/builder.rs
@@ -384,7 +384,7 @@ dead_ends: Vec::new(),
         if ret != 0 {
             return Err(std::io::Error::last_os_error());
         }
-        Ok(stat.f_bavail * stat.f_frsize)
+        Ok((stat.f_bavail as u64) * (stat.f_frsize as u64))
     }
 
     fn process_file(&mut self, path: PathBuf) -> Result<bool> {
